--- code/security/__init__.py
@@ -1,1 +1,0 @@
-from security.security import Security
