+++ code/security/security_main.py
@@ -7,7 +7,7 @@ from datetime import datetime
 import logging
 import sys
 
-from security import Security
+from security.security import Security
 
 
 def set_logger(logger_name: str = 'security', log_level: int = logging.INFO) \
